Refactor.Add some features

+++ aimgine/app/Http/Controllers/admin/home/ServiceController.php
@@ -4,8 +4,9 @@ namespace App\Http\Controllers\admin\home;
 
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
-use App\Service;
+use App\Models\Home\Service;
 use Image;
+
 class ServiceController extends Controller
 {
     public function index() {
@@ -14,11 +15,19 @@ class ServiceController extends Controller
     }
 
     public function create(){
-        return view('admin.home.service.add');
+        $services = Service::orderBy('id', 'desc')->get();
+        if($services->count() >= 4){
+            return back()->withErrors('You already have 4 services');
+        }else{
+            return view('admin.home.service.add');
+        }
     }
 
     public function store(Request $request){
-        $this->validate($request, ['text' => 'required|min:3','image' => 'required','position' => 'required|unique:services']);
+        $this->validate($request, ['text' => 'required|min:3',
+        'image' => 'required',
+        'position' => 'required|unique:home_services|digits_between:1,4']);
+
         $services = Service::orderBy('id', 'desc')->get();
         if($services->count() >= 4){
             return back()->withErrors('You already have 4 services');
@@ -32,11 +41,15 @@ class ServiceController extends Controller
                 //Used intervention image-----------------------------------//
                 try {
                   $img = Image::make($imgResource);
-                } catch (Exception $ex) {
+              } catch (\Exception $ex) {
                   return back()->withErrors('Please upload .jpeg or .png images.');
                 }
                 if ($img->filesize() < 10000000) {
-                    $img->save('./content/img/services/'.$fullImgName);
+                    try{
+                        $img->save('./content/img/services/'.$fullImgName);
+                    }catch(\Exception $e){
+                        return back()->withErrors('Cannot upload image');
+                    }
                 } else {
                     return back()->withErrors('Please upload smaller images.');
                 }
@@ -45,14 +58,14 @@ class ServiceController extends Controller
             $service->position = $request['position'];
             $service->text = $request['text'];
             $service->save();
-
-            return back()->withErrors('Service saved');
+            $services = Service::orderBy('id', 'desc')->get();
+            return view('admin.home.service.services', compact('services'))->withErrors('Service saved');
         }
 
     }
 
     public function show(Service $service){
-        return view('admin.home.service', compact('slider'));
+
     }
 
     public function edit(Service $service){
@@ -60,7 +73,8 @@ class ServiceController extends Controller
     }
 
     public function update(Request $request,Service $service){
-        $this->validate($request, ['text' => 'required|min:3','position' => 'required']);
+        $this->validate($request, ['text' => 'required|min:3',
+        'position' => 'required|unique:home_services,position,'.$service->id.',id|digits_between:1,4']);
         if($request->file('image') !== null){
             $imgResource = $request->file('image');
             $imgName = $imgResource->getClientOriginalName();
@@ -73,12 +87,22 @@ class ServiceController extends Controller
               return back()->withErrors('Please upload .jpeg or .png images.');
             }
             if ($img->filesize() < 10000000) {
-                $img->save('./content/img/services/'.$fullImgName);
+                try{
+                    $img->save('./content/img/services/'.$fullImgName);
+                }catch(\Exception $e){
+                    return back()->withErrors('Cannot upload image');
+                }
             } else {
                 return back()->withErrors('Please upload smaller images.');
             }
-            unlink('./content/img/services/' . $service->svg);
-            $service->svg = $fullImgName;
+            $fileExists = file_exists('./content/img/services/' . $service->svg);
+            if($fileExists){
+                unlink('./content/img/services/' . $service->svg);
+                $service->svg = $fullImgName;
+            }else{
+                return back()->withErrors('Service image not found');
+            }
+
         }
         $service->position = $request['position'];
         $service->text = $request['text'];
@@ -87,12 +111,12 @@ class ServiceController extends Controller
     }
 
     public function destroy(Request $request, Service $service) {
-        $service = Service::find($service->id);
         $fileExists = file_exists('./content/img/services/' . $service->svg);
-        if($service->svg != ""){
+        if($service->svg != "" && $fileExists){
             unlink('./content/img/services/' . $service->svg);
+            $service->delete();
+            return back()->withErrors('Service deleted');
         }
-        $service->delete();
-        return redirect()->back()->withErrors('Service deleted');
+        return redirect()->back()->withErrors('Something went wrong');
     }
 }
