+++ aimgine/app/Http/Controllers/admin/home/SliderController.php
@@ -4,36 +4,43 @@ namespace App\Http\Controllers\admin\home;
 
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
-use App\Slider;
+use App\Models\Home\Slider;
 use Image;
 
 class SliderController extends Controller
 {
     public function index() {
-        $sliders = Slider::orderBy('id', 'desc')->paginate(20);
+        $sliders = Slider::orderBy('id', 'desc')->paginate(5);
         return view('admin.home.slider.sliders', compact('sliders'));
     }
 
     public function create(){
-        return view('admin.home.slider.addSlider');
+        return view('admin.home.slider.add');
     }
 
     public function store(Request $request){
-        $this->validate($request, ['text' => 'required|min:3','image' => 'required','position' => 'required|unique:sliders']);
+        $this->validate($request, ['text' => 'required|min:3',
+        'image' => 'required',
+        'position' => 'required|unique:home_sliders|digits_between:1,4']);
+
         $slider = new Slider();
         if($request->file('image') !== null){
             $imgResource = $request->file('image');
             $imgName = $imgResource->getClientOriginalName();
             $randNum = rand();
             $fullImgName = $randNum.'_'.$imgName;
-            //Used intervention image-----------------------------------//
+            // //Used intervention image-----------------------------------//
             try {
               $img = Image::make($imgResource);
-            } catch (Exception $ex) {
+            } catch (\Exception $ex) {
               return back()->withErrors('Please upload .jpeg or .png images.');
             }
             if ($img->filesize() < 10000000) {
-                $img->resize(800, 470)->save('./content/img/sliders/'.$fullImgName);
+                try{
+                    $img->resize(1400, 430)->save('./content/img/sliders/'.$fullImgName);
+                }catch(\Exception $e){
+                    return back()->withErrors('Cannot upload image');
+                }
             } else {
                 return back()->withErrors('Please upload smaller images.');
             }
@@ -47,7 +54,7 @@ class SliderController extends Controller
     }
 
     public function show(Slider $slider){
-        return view('admin.home.slider', compact('slider'));
+
     }
 
     public function edit(Slider $slider){
@@ -55,7 +62,8 @@ class SliderController extends Controller
     }
 
     public function update(Request $request,Slider $slider){
-        $this->validate($request, ['text' => 'required|min:3','position' => 'required']);
+        $this->validate($request, ['text' => 'required|min:3',
+        'position' => 'required|unique:home_sliders,position,'.$slider->id.',id|digits_between:1,4']);
         if($request->file('image') !== null){
             $imgResource = $request->file('image');
             $imgName = $imgResource->getClientOriginalName();
@@ -64,16 +72,26 @@ class SliderController extends Controller
             //Used intervention image-----------------------------------//
             try {
               $img = Image::make($imgResource);
-            } catch (Exception $ex) {
+            } catch (\Exception $ex) {
               return back()->withErrors('Please upload .jpeg or .png images.');
             }
             if ($img->filesize() < 10000000) {
-                $img->resize(800, 470)->save('./content/img/sliders/'.$fullImgName);
+                try{
+                    $img->resize(1400, 430)->save('./content/img/sliders/'.$fullImgName);
+                }catch(\Exception $e){
+                    return back()->withErrors('Cannot upload image');
+                }
             } else {
                 return back()->withErrors('Please upload smaller images.');
             }
-            unlink('./content/img/sliders/' . $slider->image);
-            $slider->image = $fullImgName;
+            $fileExists = file_exists('./content/img/sliders/' . $slider->image);
+            if($fileExists){
+                unlink('./content/img/sliders/' . $slider->image);
+                $slider->image = $fullImgName;
+            }else{
+                return back()->withErrors('Slider image not found');
+            }
+
         }
         $slider->position = $request['position'];
         $slider->text = $request['text'];
@@ -82,12 +100,12 @@ class SliderController extends Controller
     }
 
     public function destroy(Request $request, Slider $slider) {
-        $slider = Slider::find($slider->id);
         $fileExists = file_exists('./content/img/sliders/' . $slider->image);
-        if($slider->image != ""){
+        if($slider->image != "" && $fileExists){
             unlink('./content/img/sliders/' . $slider->image);
+            $slider->delete();
+            return back()->withErrors('Slider deleted');
         }
-        $slider->delete();
-        return redirect()->back()->withErrors('Slider deleted');
+        return back()->withErrors('Something went wrong');
     }
 }
