+++ aimgine/app/Http/Controllers/admin/whoarewe/TeamController.php
@@ -5,7 +5,7 @@ namespace App\Http\Controllers\admin\whoarewe;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use Image;
-use App\Team;
+use App\Models\Whoarewe\Team;
 
 class TeamController extends Controller
 {
@@ -19,40 +19,41 @@ class TeamController extends Controller
     }
 
     public function store(Request $request){
-        $this->validate($request, ['text' => 'required|min:3','image' => 'required','name' => 'required']);
-        $team = Team::orderBy('id', 'desc')->get();
+        $this->validate($request, ['text' => 'required|min:3',
+        'image' => 'required',
+        'name' => 'required']);
 
-            $team = new Team();
-            if($request->file('image') !== null){
-                $imgResource = $request->file('image');
-                $imgName = $imgResource->getClientOriginalName();
-                $randNum = rand();
-                $fullImgName = $randNum.'_'.$imgName;
-                //Used intervention image-----------------------------------//
-                try {
-                  $img = Image::make($imgResource);
-                } catch (Exception $ex) {
-                  return back()->withErrors('Please upload .jpeg or .png images.');
-                }
-                if ($img->filesize() < 10000000) {
-                    $img->save('./content/img/team/'.$fullImgName);
-                } else {
-                    return back()->withErrors('Please upload smaller images.');
-                }
-                $team->image = $fullImgName;
+        $team = new Team();
+        if($request->file('image') !== null){
+            $imgResource = $request->file('image');
+            $imgName = $imgResource->getClientOriginalName();
+            $randNum = rand();
+            $fullImgName = $randNum.'_'.$imgName;
+            //Used intervention image-----------------------------------//
+            try {
+              $img = Image::make($imgResource);
+            } catch (\Exception $ex) {
+              return back()->withErrors('Please upload .jpeg or .png images.');
             }
-            $team->name = $request['name'];
-            $team->description = $request['text'];
-            $team->save();
+            if ($img->filesize() < 10000000) {
+                try {
+                    $img->save('./content/img/team/'.$fullImgName);
+                } catch (\Exception $e) {
+                    return back()->withErrors('Something went wrong');
+                }
 
-            return back()->withErrors('Team saved');
+            } else {
+                return back()->withErrors('Please upload smaller images.');
+            }
+            $team->image = $fullImgName;
+        }
+        $team->name = $request['name'];
+        $team->description = $request['text'];
+        $team->save();
 
-
+        return back()->withErrors('Team saved');
     }
 
-    public function show(Team $team){
-        return view('admin.whoarewe.team', compact('team'));
-    }
 
     public function edit(Team $team){
         return view('admin.whoarewe.team.edit', compact('team'));
@@ -68,17 +69,27 @@ class TeamController extends Controller
             //Used intervention image-----------------------------------//
             try {
               $img = Image::make($imgResource);
-            } catch (Exception $ex) {
+            } catch (\Exception $ex) {
               return back()->withErrors('Please upload .jpeg or .png images.');
             }
             if ($img->filesize() < 10000000) {
-                $img->save('./content/img/services/'.$fullImgName);
+                try {
+                    $img->save('./content/img/team/'.$fullImgName);
+                } catch (\Exception $e) {
+                    return back()->withErrors('Something went wrong');
+                }
             } else {
                 return back()->withErrors('Please upload smaller images.');
             }
-            unlink('./content/img/team/' . $team->image);
-            $team->image = $fullImgName;
+            $fileExists = file_exists('./content/img/team/' . $team->image);
+            if($fileExists){
+                unlink('./content/img/team/' . $team->image);
+                $team->image = $fullImgName;
+            }else{
+                return back()->withErrors('Something went wrong');
+            }
         }
+
         $team->name = $request['name'];
         $team->description = $request['text'];
         $team->save();
@@ -86,14 +97,14 @@ class TeamController extends Controller
     }
 
     public function destroy(Request $request, Team $team) {
-
-
-        // $team = Team::find($team->id);
         $fileExists = file_exists('./content/img/team/' . $team->image);
-        if($team->image != ""){
+        if($team->image != "" && $fileExists){
             unlink('./content/img/team/' . $team->image);
+            $team->delete();
+            return redirect()->back()->withErrors('Team deleted');
+        }else{
+            return redirect()->back()->withErrors('Something went wrong');
         }
-        $team->delete();
-        return redirect()->back()->withErrors('Team deleted');
+
     }
 }
